Refactor classes and Create more sub-packages.
The application only has one main package and does not provide a good
abstraction on functionality of different classes.

Let's add some sub-packages to arrange classes according to their
functionality.

Now functionality of different classes are clearly shown.
