+++ src/main/java/windycall/Handler/AddEventHandler.java
@@ -1,4 +1,8 @@
-package windycall;
+package windycall.Handler;
+
+import windycall.task.Event;
+import windycall.task.Task;
+import windycall.exception.WindyCallException;
 
 public class AddEventHandler extends AddTaskHandler {
