+++ src/main/java/windycall/parser/Parser.java
@@ -1,8 +1,13 @@
-package windycall;
+package windycall.parser;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
 import javafx.util.Pair;
+import windycall.Handler.AddDeadlineHandler;
+import windycall.Handler.AddEventHandler;
+import windycall.Handler.AddTodoHandler;
+import windycall.OperationType;
+import windycall.exception.WindyCallException;
 
 /**
  * Deals with making sense of the user command by splitting
@@ -100,6 +105,24 @@ public class Parser {
             return new Pair<>(-1, "☹ OOPS!!! You should input a number");
         }
         return new Pair<>(num, "Success");
+    }
+
+    public OperationType getAddTaskType(String message) throws WindyCallException {
+        String[] parts = message.split(" ");
+        switch (parts[0]) {
+        case "todo":
+            return OperationType.TODO;
+//            break;
+        case "deadline":
+            return OperationType.DEADLINE;
+//            break;
+        case "event":
+            return OperationType.EVENT;
+//            break;
+        default:
+            throw new WindyCallException("☹ OOPS!!! I'm sorry, but I don't know what that means :-(");
+//          break;
+        }
     }
 
     /**
