+++ src/main/java/windycall/storage/Storage.java
@@ -1,4 +1,10 @@
-package windycall;
+package windycall.storage;
+
+import windycall.ui.Ui;
+import windycall.task.Deadline;
+import windycall.task.Event;
+import windycall.task.Task;
+import windycall.task.Todo;
 
 import java.io.File;
 import java.io.FileNotFoundException;
