+++ src/main/java/windycall/task/TaskList.java
@@ -1,4 +1,4 @@
-package windycall;
+package windycall.task;
 
 public class TaskList {
 }
